apply low-level - generation principle on I2S peripheral
