+++ esp-hal/src/i2s/master/mod.rs
@@ -148,10 +148,13 @@ use core::mem::ManuallyDrop;
 use enumset::{EnumSet, EnumSetType, enum_set};
 use private::*;
 
+mod low_level;
+
+#[doc(hidden)]
+pub use low_level::Info;
+
 #[cfg(any(i2s_supports_pdm_tx, i2s_supports_pdm_rx))]
 pub use super::pdm::{PdmConfig, PdmError, PdmInstance, PdmRxConfig, PdmSlotMode, PdmTxConfig};
-#[cfg(i2s_version = "1")]
-use crate::RegisterToggle;
 use crate::{
     Async,
     Blocking,
@@ -1805,14 +1808,6 @@ pub(crate) mod private {
         pub(crate) numerator: u32,
     }
 
-    #[cfg(not(i2s_version = "1"))]
-    pub struct I2sMclkDividers {
-        pub(crate) x: u32,
-        pub(crate) y: u32,
-        pub(crate) z: u32,
-        pub(crate) yn1: bool,
-    }
-
     impl I2sClockDividers {
         pub fn new(sample_rate: Rate, channels: u8, data_bits: u8) -> I2sClockDividers {
             // this loosely corresponds to `i2s_std_calculate_clock` and
@@ -1876,874 +1871,5 @@ pub(crate) mod private {
                 numerator,
             }
         }
-
-        #[cfg(not(i2s_version = "1"))]
-        pub(crate) fn mclk_dividers(&self) -> I2sMclkDividers {
-            let (x, y, z, yn1) = if self.denominator == 0 || self.numerator == 0 {
-                (0, 0, 0, true)
-            } else if self.numerator > self.denominator / 2 {
-                let x = self
-                    .denominator
-                    .overflowing_div(self.denominator.overflowing_sub(self.numerator).0)
-                    .0
-                    .overflowing_sub(1)
-                    .0;
-                let y = self.denominator % (self.denominator.overflowing_sub(self.numerator).0);
-                let z = self.denominator.overflowing_sub(self.numerator).0;
-                (x, y, z, true)
-            } else {
-                let x = self.denominator / self.numerator - 1;
-                let y = self.denominator % self.numerator;
-                let z = self.numerator;
-                (x, y, z, false)
-            };
-
-            I2sMclkDividers { x, y, z, yn1 }
-        }
-    }
-}
-
-/// Peripheral data describing a particular I2S instance.
-///
-/// All the per-instance data (register block pointer, system peripheral marker and the
-/// GPIO matrix signals) is stored here, so that the driver can operate on a single
-/// type-erased `&'static Info` regardless of which concrete I2S peripheral is used.
-#[doc(hidden)]
-#[non_exhaustive]
-pub struct Info {
-    /// Pointer to the register block for this I2S instance.
-    pub register_block: *const RegisterBlock,
-
-    /// The system peripheral marker.
-    pub peripheral: crate::system::Peripheral,
-
-    /// MCLK output signal.
-    #[cfg(not(esp32))] // MCLK on ESP32 requires special handling
-    pub mclk: OutputSignal,
-
-    /// BCLK (TX) output signal.
-    pub bclk: OutputSignal,
-
-    /// WS (TX) output signal.
-    pub ws: OutputSignal,
-
-    /// BCLK (RX) output signal.
-    pub bclk_rx: OutputSignal,
-
-    /// WS (RX) output signal.
-    pub ws_rx: OutputSignal,
-
-    /// Data out signals.
-    pub dout_lines: &'static [OutputSignal],
-
-    /// Data in signals.
-    pub din_lines: &'static [InputSignal],
-
-    /// PDM TX supported.
-    pub pdm_tx: bool,
-
-    /// PDM RX supported.
-    pub pdm_rx: bool,
-
-    /// Hardware PCM-to-PDM format conversion filter supported on TX.
-    pub pcm2pdm: bool,
-
-    /// Hardware PDM-to-PCM format conversion filter supported on RX.
-    pub pdm2pcm: bool,
-}
-
-// SAFETY: The register block pointer refers to a static peripheral memory region.
-unsafe impl Sync for Info {}
-
-impl Info {
-    pub(crate) fn regs(&self) -> &RegisterBlock {
-        unsafe { &*self.register_block }
-    }
-
-    fn dout(&self, line: u8) -> Option<OutputSignal> {
-        // (line 1 for two-line DAC).
-        self.dout_lines.get(line as usize).copied()
-    }
-
-    fn din(&self, line: u8) -> Option<InputSignal> {
-        // (line 0 is the default DIN signal).
-        self.din_lines.get(line as usize).copied()
-    }
-
-    fn interrupts(&self) -> EnumSet<I2sInterrupt> {
-        let mut res = EnumSet::new();
-        let ints = self.regs().int_st().read();
-
-        if ints.rx_hung().bit() {
-            res.insert(I2sInterrupt::RxHung);
-        }
-        if ints.tx_hung().bit() {
-            res.insert(I2sInterrupt::TxHung);
-        }
-        #[cfg(not(i2s_version = "1"))]
-        if ints.rx_done().bit() {
-            res.insert(I2sInterrupt::RxDone);
-        }
-        #[cfg(not(i2s_version = "1"))]
-        if ints.tx_done().bit() {
-            res.insert(I2sInterrupt::TxDone);
-        }
-
-        res
-    }
-
-    fn enable_listen(&self, interrupts: EnumSet<I2sInterrupt>, enable: bool) {
-        self.regs().int_ena().modify(|_, w| {
-            for interrupt in interrupts {
-                match interrupt {
-                    I2sInterrupt::RxHung => w.rx_hung().bit(enable),
-                    I2sInterrupt::TxHung => w.tx_hung().bit(enable),
-                    #[cfg(not(i2s_version = "1"))]
-                    I2sInterrupt::RxDone => w.rx_done().bit(enable),
-                    #[cfg(not(i2s_version = "1"))]
-                    I2sInterrupt::TxDone => w.tx_done().bit(enable),
-                };
-            }
-            w
-        });
-    }
-
-    fn clear_interrupts(&self, interrupts: EnumSet<I2sInterrupt>) {
-        self.regs().int_clr().write(|w| {
-            for interrupt in interrupts {
-                match interrupt {
-                    I2sInterrupt::RxHung => w.rx_hung().clear_bit_by_one(),
-                    I2sInterrupt::TxHung => w.tx_hung().clear_bit_by_one(),
-                    #[cfg(not(i2s_version = "1"))]
-                    I2sInterrupt::RxDone => w.rx_done().clear_bit_by_one(),
-                    #[cfg(not(i2s_version = "1"))]
-                    I2sInterrupt::TxDone => w.tx_done().clear_bit_by_one(),
-                };
-            }
-            w
-        });
-    }
-}
-
-#[cfg(i2s_version = "1")]
-impl Info {
-    pub(crate) fn set_clock(&self, clock_settings: I2sClockDividers) {
-        self.regs().clkm_conf().modify(|r, w| unsafe {
-            // select PLL_160M
-            w.bits(r.bits() | (property!("i2s.default_clock_source") << 21))
-        });
-
-        #[cfg(esp32)]
-        self.regs()
-            .clkm_conf()
-            .modify(|_, w| w.clka_ena().clear_bit());
-
-        self.regs().clkm_conf().modify(|_, w| unsafe {
-            w.clk_en().set_bit();
-            w.clkm_div_num().bits(clock_settings.mclk_divider as u8);
-            w.clkm_div_a().bits(clock_settings.denominator as u8);
-            w.clkm_div_b().bits(clock_settings.numerator as u8)
-        });
-
-        self.regs().sample_rate_conf().modify(|_, w| unsafe {
-            w.tx_bck_div_num().bits(clock_settings.bclk_divider as u8);
-            w.rx_bck_div_num().bits(clock_settings.bclk_divider as u8)
-        });
-    }
-
-    fn update_tx(&self) {
-        // Nothing to do.
-    }
-
-    fn update_rx(&self) {
-        // Nothing to do.
-    }
-
-    fn configure(&self, config: &Config) -> Result<(), ConfigError> {
-        config.validate(self)?;
-
-        match config {
-            Config::Tdm(c) => {
-                self.configure_tx(&c.tx_config, c.data_format)?;
-                self.configure_rx(&c.rx_config, c.data_format)?;
-
-                self.set_clock(config.calculate_clock());
-
-                self.regs().sample_rate_conf().modify(|_, w| unsafe {
-                    // Having different data formats for each direction would make clock
-                    // calculations more tricky
-                    w.tx_bits_mod().bits(c.data_format.data_bits());
-                    w.rx_bits_mod().bits(c.data_format.data_bits())
-                });
-
-                self.regs().conf().modify(|_, w| {
-                    w.tx_slave_mod().clear_bit();
-                    w.rx_slave_mod().bit(c.signal_loopback);
-                    // Send MSB to the right channel to be consistent with ESP32-S3 et al.
-                    w.tx_msb_right().set_bit();
-                    w.rx_msb_right().set_bit();
-                    // ESP32 generates two clock pulses first. If the WS is low, those first
-                    // clock pulses are indistinguishable from real
-                    // data, which corrupts the first few samples. So we
-                    // send the right channel first (which means WS is high during
-                    // the first sample) to prevent this issue.
-                    w.tx_right_first().set_bit();
-                    w.rx_right_first().set_bit();
-                    w.tx_mono().clear_bit();
-                    w.rx_mono().clear_bit();
-                    w.sig_loopback().bit(c.signal_loopback)
-                });
-
-                self.regs().fifo_conf().modify(|_, w| w.dscr_en().set_bit());
-
-                self.regs().conf1().modify(|_, w| {
-                    w.tx_pcm_bypass().set_bit();
-                    w.rx_pcm_bypass().set_bit()
-                });
-
-                self.regs().pd_conf().modify(|_, w| {
-                    w.fifo_force_pu().set_bit();
-                    w.fifo_force_pd().clear_bit()
-                });
-
-                self.regs().conf2().modify(|_, w| {
-                    w.camera_en().clear_bit();
-                    w.lcd_en().clear_bit()
-                });
-            }
-            #[cfg(any(i2s_supports_pdm_tx, i2s_supports_pdm_rx))]
-            Config::Pdm(c) => {
-                crate::i2s::pdm::configure_pdm(self, c)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn configure_tx(
-        &self,
-        config: &UnitConfig,
-        data_format: DataFormat,
-    ) -> Result<(), ConfigError> {
-        config.validate()?;
-
-        let chan_mod = match config.channels {
-            Channels::STEREO | Channels::MONO => 0,
-            Channels::LEFT => 3,
-            Channels::RIGHT => 4,
-            _ => unreachable!(),
-        };
-
-        let fifo_mod = match (data_format.data_bits(), config.channels == Channels::STEREO) {
-            (8 | 16, true) => 0,
-            (8 | 16, false) => 1,
-            (24 | 32, true) => 2,
-            (24 | 32, false) => 3,
-            _ => unreachable!(),
-        };
-
-        self.regs().conf().modify(|_, w| {
-            w.tx_msb_shift().bit(config.msb_shift);
-            // Short frame synchronization
-            w.tx_short_sync().bit(config.ws_width == WsWidth::Bit)
-        });
-
-        self.regs().conf1().modify(|_, w| w.tx_stop_en().set_bit());
-
-        #[cfg(not(esp32))]
-        self.regs().conf().modify(|_, w| {
-            // Channel configurations other than Stereo should use same data from DMA
-            // for both channels
-            w.tx_dma_equal().bit(config.channels != Channels::STEREO);
-            // Byte endianness
-            w.tx_big_endian()
-                .bit(config.endianness == Endianness::BigEndian)
-        });
-
-        self.regs().fifo_conf().modify(|_, w| unsafe {
-            w.tx_fifo_mod().bits(fifo_mod);
-            w.tx_fifo_mod_force_en().set_bit()
-        });
-
-        self.regs()
-            .conf_sigle_data()
-            .modify(|_, w| unsafe { w.sigle_data().bits(config.channels.fill.unwrap_or(0)) });
-
-        self.regs()
-            .conf_chan()
-            .modify(|_, w| unsafe { w.tx_chan_mod().bits(chan_mod) });
-
-        Ok(())
-    }
-
-    fn configure_rx(
-        &self,
-        config: &UnitConfig,
-        data_format: DataFormat,
-    ) -> Result<(), ConfigError> {
-        config.validate()?;
-
-        let chan_mod = match config.channels {
-            Channels::STEREO => 0,
-            Channels::LEFT | Channels::MONO => 1,
-            Channels::RIGHT => 2,
-            _ => unreachable!(),
-        };
-
-        let fifo_mod = match (data_format.data_bits(), config.channels == Channels::STEREO) {
-            (8 | 16, true) => 0,
-            (8 | 16, false) => 1,
-            (24 | 32, true) => 2,
-            (24 | 32, false) => 3,
-            _ => unreachable!(),
-        };
-
-        self.regs().conf().modify(|_, w| {
-            w.rx_msb_shift().bit(config.msb_shift);
-            // Short frame synchronization
-            w.rx_short_sync().bit(config.ws_width == WsWidth::Bit)
-        });
-
-        #[cfg(not(esp32))]
-        self.regs().conf().modify(|_, w| {
-            // Channel configurations other than Stereo should use same data from DMA
-            // for both channels
-            w.rx_dma_equal().bit(config.channels != Channels::STEREO);
-            // Byte endianness
-            w.rx_big_endian()
-                .bit(config.endianness == Endianness::BigEndian)
-        });
-
-        self.regs().fifo_conf().modify(|_, w| unsafe {
-            w.rx_fifo_mod().bits(fifo_mod);
-            w.rx_fifo_mod_force_en().set_bit()
-        });
-
-        self.regs()
-            .conf_chan()
-            .modify(|_, w| unsafe { w.rx_chan_mod().bits(chan_mod) });
-
-        Ok(())
-    }
-
-    fn set_master(&self) {
-        self.regs().conf().modify(|_, w| {
-            w.rx_slave_mod().clear_bit();
-            w.tx_slave_mod().clear_bit()
-        });
-    }
-
-    fn reset_tx(&self) {
-        self.regs().conf().modify(|_, w| w.tx_reset().bit(true));
-        self.regs()
-            .conf()
-            .modify(|_, w| w.tx_fifo_reset().bit(true));
-
-        self.regs().conf().modify(|_, w| {
-            w.tx_reset().bit(false);
-            w.tx_fifo_reset().bit(false)
-        });
-
-        #[cfg(esp32s2)]
-        while self.regs().conf().read().tx_reset_st().bit_is_set() {}
-
-        #[cfg(esp32)]
-        while self.regs().state().read().tx_fifo_reset_back().bit_is_set() {}
-
-        self.regs().lc_conf().modify(|_, w| w.out_rst().bit(true));
-        self.regs().lc_conf().modify(|_, w| w.out_rst().bit(false));
-
-        self.regs().int_clr().write(|w| {
-            w.out_done().clear_bit_by_one();
-            w.out_total_eof().clear_bit_by_one()
-        });
-    }
-
-    fn tx_start(&self) {
-        self.regs().conf().modify(|_, w| w.tx_start().set_bit());
-
-        while self.regs().state().read().tx_idle().bit_is_set() {
-            // wait
-        }
-    }
-
-    fn tx_stop(&self) {
-        self.regs().conf().modify(|_, w| w.tx_start().clear_bit());
-    }
-
-    fn is_tx_done(&self) -> bool {
-        self.regs().state().read().tx_idle().bit_is_set()
-    }
-
-    fn reset_rx(&self) {
-        self.regs().conf().toggle(|w, bit| {
-            w.rx_reset().bit(bit);
-            w.rx_fifo_reset().bit(bit)
-        });
-
-        #[cfg(esp32s2)]
-        while self.regs().conf().read().rx_reset_st().bit_is_set() {}
-
-        self.regs().lc_conf().toggle(|w, bit| w.in_rst().bit(bit));
-
-        self.regs().int_clr().write(|w| {
-            w.in_done().clear_bit_by_one();
-            w.in_suc_eof().clear_bit_by_one()
-        });
-    }
-
-    fn rx_start(&self, len: usize) {
-        self.regs()
-            .int_clr()
-            .write(|w| w.in_suc_eof().clear_bit_by_one());
-
-        let eof_num = cfg_select! {
-            // On ESP32, the eof_num count in words.
-            esp32 => len / 4,
-            _ => len - 1,
-        };
-
-        self.regs()
-            .rxeof_num()
-            .modify(|_, w| unsafe { w.rx_eof_num().bits(eof_num as u32) });
-
-        self.regs().conf().modify(|_, w| w.rx_start().set_bit());
-    }
-
-    fn rx_stop(&self) {
-        self.regs().conf().modify(|_, w| w.rx_start().clear_bit());
-    }
-
-    fn is_rx_done(&self) -> bool {
-        self.regs().int_raw().read().in_dscr_empty().bit_is_set()
-    }
-}
-
-#[cfg(not(i2s_version = "1"))]
-impl Info {
-    #[cfg(not(any(i2s_clock_configured_by_pcr, i2s_clock_configured_by_hp_sys_clkrst)))]
-    pub(crate) fn set_tx_clock(&self, clock_settings: I2sClockDividers) {
-        let clkm_div = clock_settings.mclk_dividers();
-
-        self.regs().tx_clkm_div_conf().modify(|_, w| unsafe {
-            w.tx_clkm_div_x().bits(clkm_div.x as u16);
-            w.tx_clkm_div_y().bits(clkm_div.y as u16);
-            w.tx_clkm_div_yn1().bit(clkm_div.yn1);
-            w.tx_clkm_div_z().bits(clkm_div.z as u16)
-        });
-
-        self.regs().tx_clkm_conf().modify(|_, w| unsafe {
-            w.clk_en().set_bit();
-            w.tx_clk_active().set_bit();
-            // for now fixed at 160MHz
-            w.tx_clk_sel().bits(property!("i2s.default_clock_source"));
-            w.tx_clkm_div_num().bits(clock_settings.mclk_divider as u8)
-        });
-
-        #[cfg(i2s_version = "2")]
-        self.regs().tx_conf1().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-        #[cfg(i2s_version = "3")]
-        self.regs().tx_conf().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    #[cfg(not(any(i2s_clock_configured_by_pcr, i2s_clock_configured_by_hp_sys_clkrst)))]
-    pub(crate) fn set_rx_clock(&self, clock_settings: I2sClockDividers) {
-        let clkm_div = clock_settings.mclk_dividers();
-
-        self.regs().rx_clkm_div_conf().modify(|_, w| unsafe {
-            w.rx_clkm_div_x().bits(clkm_div.x as u16);
-            w.rx_clkm_div_y().bits(clkm_div.y as u16);
-            w.rx_clkm_div_yn1().bit(clkm_div.yn1);
-            w.rx_clkm_div_z().bits(clkm_div.z as u16)
-        });
-
-        self.regs().rx_clkm_conf().modify(|_, w| unsafe {
-            w.rx_clk_active().set_bit();
-            // for now fixed at 160MHz
-            w.rx_clk_sel().bits(property!("i2s.default_clock_source"));
-            w.rx_clkm_div_num().bits(clock_settings.mclk_divider as u8);
-            w.mclk_sel().bit(true)
-        });
-
-        #[cfg(i2s_version = "2")]
-        self.regs().rx_conf1().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-        #[cfg(i2s_version = "3")]
-        self.regs().rx_conf().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    #[cfg(i2s_clock_configured_by_pcr)]
-    pub(crate) fn set_tx_clock(&self, clock_settings: I2sClockDividers) {
-        // I2S clocks are configured via PCR
-        use crate::peripherals::PCR;
-
-        let clkm_div = clock_settings.mclk_dividers();
-        let pcr = PCR::regs();
-
-        // Pulse a temporary divider before applying the target coefficients to avoid
-        // a hardware glitch where the clock divider applies twice on PCR chips.
-        pcr.i2s_tx_clkm_conf()
-            .modify(|_, w| unsafe { w.i2s_tx_clkm_div_num().bits(2) });
-        pcr.i2s_tx_clkm_div_conf().modify(|_, w| unsafe {
-            w.i2s_tx_clkm_div_yn1().clear_bit();
-            w.i2s_tx_clkm_div_y().bits(1);
-            w.i2s_tx_clkm_div_z().bits(0);
-            w.i2s_tx_clkm_div_x().bits(0)
-        });
-
-        pcr.i2s_tx_clkm_div_conf().modify(|_, w| unsafe {
-            w.i2s_tx_clkm_div_x().bits(clkm_div.x as u16);
-            w.i2s_tx_clkm_div_y().bits(clkm_div.y as u16);
-            w.i2s_tx_clkm_div_yn1().bit(clkm_div.yn1);
-            w.i2s_tx_clkm_div_z().bits(clkm_div.z as u16)
-        });
-
-        pcr.i2s_tx_clkm_conf().modify(|_, w| unsafe {
-            w.i2s_tx_clkm_en().set_bit();
-            // for now fixed at 160MHz for C6 and 96MHz for H2
-            w.i2s_tx_clkm_sel()
-                .bits(property!("i2s.default_clock_source"));
-            w.i2s_tx_clkm_div_num()
-                .bits(clock_settings.mclk_divider as u8)
-        });
-
-        #[cfg(i2s_version = "2")]
-        self.regs().tx_conf1().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-
-        #[cfg(i2s_version = "3")]
-        self.regs().tx_conf().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    #[cfg(i2s_clock_configured_by_pcr)]
-    pub(crate) fn set_rx_clock(&self, clock_settings: I2sClockDividers) {
-        // I2S clocks are configured via PCR
-        use crate::peripherals::PCR;
-
-        let clkm_div = clock_settings.mclk_dividers();
-        let pcr = PCR::regs();
-
-        // Pulse a temporary divider before applying the target coefficients to avoid
-        // a hardware glitch where the clock divider applies twice on PCR chips.
-        pcr.i2s_rx_clkm_conf()
-            .modify(|_, w| unsafe { w.i2s_rx_clkm_div_num().bits(2) });
-        pcr.i2s_rx_clkm_div_conf().modify(|_, w| unsafe {
-            w.i2s_rx_clkm_div_yn1().clear_bit();
-            w.i2s_rx_clkm_div_y().bits(1);
-            w.i2s_rx_clkm_div_z().bits(0);
-            w.i2s_rx_clkm_div_x().bits(0)
-        });
-
-        pcr.i2s_rx_clkm_div_conf().modify(|_, w| unsafe {
-            w.i2s_rx_clkm_div_x().bits(clkm_div.x as u16);
-            w.i2s_rx_clkm_div_y().bits(clkm_div.y as u16);
-            w.i2s_rx_clkm_div_yn1().bit(clkm_div.yn1);
-            w.i2s_rx_clkm_div_z().bits(clkm_div.z as u16)
-        });
-
-        pcr.i2s_rx_clkm_conf().modify(|_, w| unsafe {
-            w.i2s_rx_clkm_en().set_bit();
-            // for now fixed at 160MHz for C6 and 96MHz for H2
-            w.i2s_rx_clkm_sel()
-                .bits(property!("i2s.default_clock_source"));
-            w.i2s_rx_clkm_div_num()
-                .bits(clock_settings.mclk_divider as u8);
-            w.i2s_mclk_sel().bit(true)
-        });
-
-        #[cfg(i2s_version = "2")]
-        self.regs().rx_conf1().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-
-        #[cfg(i2s_version = "3")]
-        self.regs().rx_conf().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    #[cfg(i2s_clock_configured_by_hp_sys_clkrst)]
-    pub(crate) fn set_tx_clock(&self, clock_settings: I2sClockDividers) {
-        super::hp_sys_clkrst::set_tx_clock(self.peripheral, &clock_settings);
-
-        #[cfg(i2s_version = "2")]
-        self.regs().tx_conf1().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-
-        #[cfg(i2s_version = "3")]
-        self.regs().tx_conf().modify(|_, w| unsafe {
-            w.tx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    #[cfg(i2s_clock_configured_by_hp_sys_clkrst)]
-    pub(crate) fn set_rx_clock(&self, clock_settings: I2sClockDividers) {
-        super::hp_sys_clkrst::set_rx_clock(self.peripheral, &clock_settings);
-
-        #[cfg(i2s_version = "2")]
-        self.regs().rx_conf1().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-
-        #[cfg(i2s_version = "3")]
-        self.regs().rx_conf().modify(|_, w| unsafe {
-            w.rx_bck_div_num()
-                .bits((clock_settings.bclk_divider - 1) as u8)
-        });
-    }
-
-    fn update_tx(&self) {
-        self.regs().tx_conf().modify(|_, w| w.tx_update().set_bit());
-        while self.regs().tx_conf().read().tx_update().bit_is_set() {
-            // wait
-        }
-    }
-
-    fn update_rx(&self) {
-        self.regs().rx_conf().modify(|_, w| w.rx_update().set_bit());
-        while self.regs().rx_conf().read().rx_update().bit_is_set() {
-            // wait
-        }
-    }
-
-    fn configure(&self, config: &Config) -> Result<(), ConfigError> {
-        config.validate(self)?;
-
-        match config {
-            Config::Tdm(c) => {
-                self.configure_tx(&c.tx_config)?;
-                self.configure_rx(&c.rx_config)?;
-
-                self.regs()
-                    .tx_conf()
-                    .modify(|_, w| w.sig_loopback().bit(c.signal_loopback));
-                self.regs()
-                    .rx_conf()
-                    .modify(|_, w| w.rx_slave_mod().bit(c.signal_loopback));
-            }
-            #[cfg(any(i2s_supports_pdm_tx, i2s_supports_pdm_rx))]
-            Config::Pdm(c) => {
-                crate::i2s::pdm::configure_pdm(self, c)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn configure_tx(&self, config: &UnitConfig) -> Result<(), ConfigError> {
-        use bitfield::Bit;
-
-        let ws_width = config.calculate_ws_width()?;
-        self.set_tx_clock(config.calculate_clock());
-
-        self.regs().tx_conf1().modify(|_, w| unsafe {
-            #[cfg(i2s_version = "2")]
-            w.tx_msb_shift().bit(config.msb_shift);
-            #[allow(clippy::useless_conversion)]
-            w.tx_tdm_ws_width().bits((ws_width - 1).try_into().unwrap());
-            w.tx_bits_mod().bits(config.data_format.data_bits() - 1);
-            w.tx_tdm_chan_bits()
-                .bits(config.data_format.channel_bits() - 1);
-            w.tx_half_sample_bits()
-                .bits((config.data_format.data_bits() * config.channels.count) / 2 - 1)
-        });
-
-        self.regs().tx_conf().modify(|_, w| unsafe {
-            w.tx_mono().clear_bit();
-            w.tx_mono_fst_vld().set_bit();
-            w.tx_stop_en().set_bit();
-            w.tx_chan_equal().bit(config.channels.fill.is_none());
-            w.tx_tdm_en().set_bit();
-            w.tx_pdm_en().clear_bit();
-            w.tx_pcm_bypass().set_bit();
-            #[cfg(i2s_version = "3")]
-            w.tx_msb_shift().bit(config.msb_shift);
-            w.tx_big_endian()
-                .bit(config.endianness == Endianness::BigEndian);
-            w.tx_bit_order().bit(config.bit_order == BitOrder::LsbFirst);
-            w.tx_ws_idle_pol()
-                .bit(config.ws_polarity == Polarity::ActiveHigh);
-            w.tx_chan_mod().bits(0)
-        });
-
-        self.regs().tx_tdm_ctrl().modify(|_, w| unsafe {
-            w.tx_tdm_tot_chan_num().bits(config.channels.count - 1);
-            w.tx_tdm_chan0_en().bit(config.channels.mask.bit(0));
-            w.tx_tdm_chan1_en().bit(config.channels.mask.bit(1));
-            w.tx_tdm_chan2_en().bit(config.channels.mask.bit(2));
-            w.tx_tdm_chan3_en().bit(config.channels.mask.bit(3));
-            w.tx_tdm_chan4_en().bit(config.channels.mask.bit(4));
-            w.tx_tdm_chan5_en().bit(config.channels.mask.bit(5));
-            w.tx_tdm_chan6_en().bit(config.channels.mask.bit(6));
-            w.tx_tdm_chan7_en().bit(config.channels.mask.bit(7));
-            w.tx_tdm_chan8_en().bit(config.channels.mask.bit(8));
-            w.tx_tdm_chan9_en().bit(config.channels.mask.bit(9));
-            w.tx_tdm_chan10_en().bit(config.channels.mask.bit(10));
-            w.tx_tdm_chan11_en().bit(config.channels.mask.bit(11));
-            w.tx_tdm_chan12_en().bit(config.channels.mask.bit(12));
-            w.tx_tdm_chan13_en().bit(config.channels.mask.bit(13));
-            w.tx_tdm_chan14_en().bit(config.channels.mask.bit(14));
-            w.tx_tdm_chan15_en().bit(config.channels.mask.bit(15));
-            w.tx_tdm_skip_msk_en().clear_bit()
-        });
-
-        self.regs()
-            .conf_sigle_data()
-            .modify(|_, w| unsafe { w.single_data().bits(config.channels.fill.unwrap_or(0)) });
-
-        Ok(())
-    }
-
-    fn configure_rx(&self, config: &UnitConfig) -> Result<(), ConfigError> {
-        use bitfield::Bit;
-
-        let ws_width = config.calculate_ws_width()?;
-        self.set_rx_clock(config.calculate_clock());
-
-        self.regs().rx_conf1().modify(|_, w| unsafe {
-            #[cfg(i2s_version = "2")]
-            w.rx_msb_shift().bit(config.msb_shift);
-            #[allow(clippy::useless_conversion)]
-            w.rx_tdm_ws_width().bits((ws_width - 1).try_into().unwrap());
-            w.rx_bits_mod().bits(config.data_format.data_bits() - 1);
-            w.rx_tdm_chan_bits()
-                .bits(config.data_format.channel_bits() - 1);
-            w.rx_half_sample_bits()
-                .bits((config.data_format.data_bits() * config.channels.count) / 2 - 1)
-        });
-
-        self.regs().rx_conf().modify(|_, w| unsafe {
-            w.rx_mono().clear_bit();
-            w.rx_mono_fst_vld().set_bit();
-            w.rx_stop_mode().bits(2);
-            w.rx_tdm_en().set_bit();
-            w.rx_pdm_en().clear_bit();
-            w.rx_pcm_bypass().set_bit();
-            #[cfg(i2s_version = "3")]
-            w.rx_msb_shift().bit(config.msb_shift);
-            w.rx_big_endian()
-                .bit(config.endianness == Endianness::BigEndian);
-            w.rx_bit_order().bit(config.bit_order == BitOrder::LsbFirst);
-            w.rx_ws_idle_pol()
-                .bit(config.ws_polarity == Polarity::ActiveHigh)
-        });
-
-        self.regs().rx_tdm_ctrl().modify(|_, w| unsafe {
-            w.rx_tdm_tot_chan_num().bits(config.channels.count - 1);
-            w.rx_tdm_pdm_chan0_en().bit(config.channels.mask.bit(0));
-            w.rx_tdm_pdm_chan1_en().bit(config.channels.mask.bit(1));
-            w.rx_tdm_pdm_chan2_en().bit(config.channels.mask.bit(2));
-            w.rx_tdm_pdm_chan3_en().bit(config.channels.mask.bit(3));
-            w.rx_tdm_pdm_chan4_en().bit(config.channels.mask.bit(4));
-            w.rx_tdm_pdm_chan5_en().bit(config.channels.mask.bit(5));
-            w.rx_tdm_pdm_chan6_en().bit(config.channels.mask.bit(6));
-            w.rx_tdm_pdm_chan7_en().bit(config.channels.mask.bit(7));
-            w.rx_tdm_chan8_en().bit(config.channels.mask.bit(8));
-            w.rx_tdm_chan9_en().bit(config.channels.mask.bit(9));
-            w.rx_tdm_chan10_en().bit(config.channels.mask.bit(10));
-            w.rx_tdm_chan11_en().bit(config.channels.mask.bit(11));
-            w.rx_tdm_chan12_en().bit(config.channels.mask.bit(12));
-            w.rx_tdm_chan13_en().bit(config.channels.mask.bit(13));
-            w.rx_tdm_chan14_en().bit(config.channels.mask.bit(14));
-            w.rx_tdm_chan15_en().bit(config.channels.mask.bit(15))
-        });
-
-        Ok(())
-    }
-
-    fn set_master(&self) {
-        self.regs()
-            .tx_conf()
-            .modify(|_, w| w.tx_slave_mod().clear_bit());
-        self.regs()
-            .rx_conf()
-            .modify(|_, w| w.rx_slave_mod().clear_bit());
-    }
-
-    fn reset_tx(&self) {
-        // I2S v2/v3: reset fields are write-to-trigger (WT); writing 0 has no effect.
-        self.regs().tx_conf().modify(|_, w| {
-            w.tx_reset().set_bit();
-            w.tx_fifo_reset().set_bit()
-        });
-
-        self.regs().int_clr().write(|w| {
-            w.tx_done().clear_bit_by_one();
-            w.tx_hung().clear_bit_by_one()
-        });
-    }
-
-    fn tx_start(&self) {
-        self.regs().tx_conf().modify(|_, w| w.tx_start().set_bit());
-    }
-
-    fn tx_stop(&self) {
-        self.regs()
-            .tx_conf()
-            .modify(|_, w| w.tx_start().clear_bit());
-    }
-
-    fn is_tx_done(&self) -> bool {
-        self.regs().state().read().tx_idle().bit_is_set()
-    }
-
-    fn reset_rx(&self) {
-        self.regs()
-            .rx_conf()
-            .modify(|_, w| w.rx_start().clear_bit());
-
-        // I2S v2/v3: reset fields are write-to-trigger (WT); writing 0 has no effect.
-        self.regs().rx_conf().modify(|_, w| {
-            w.rx_reset().set_bit();
-            w.rx_fifo_reset().set_bit()
-        });
-
-        self.regs().int_clr().write(|w| {
-            w.rx_done().clear_bit_by_one();
-            w.rx_hung().clear_bit_by_one()
-        });
-    }
-
-    fn rx_start(&self, len: usize) {
-        let len = len - 1;
-
-        self.regs()
-            .rxeof_num()
-            .write(|w| unsafe { w.rx_eof_num().bits(len as u16) });
-        // Sync configuration into the I2S clock domain before starting RX.
-        self.update_rx();
-        self.regs().rx_conf().modify(|_, w| w.rx_start().set_bit());
-    }
-
-    fn rx_stop(&self) {
-        self.regs()
-            .rx_conf()
-            .modify(|_, w| w.rx_start().clear_bit());
-    }
-
-    fn is_rx_done(&self) -> bool {
-        self.regs().int_raw().read().rx_done().bit_is_set()
     }
 }
